+++ httpx_auth/__init__.py
@@ -17,7 +17,7 @@ from httpx_auth.authentication import (
     OAuth2ResourceOwnerPasswordCredentials,
 )
 from httpx_auth.oauth2_tokens import JsonTokenFileCache
-from httpx_auth.aws4auth import AWS4Auth, StrictAWS4Auth
+from httpx_auth.aws import AWS4Auth, StrictAWS4Auth
 from httpx_auth.errors import (
     GrantNotProvided,
     TimeoutOccurred,
